HACK install `mkbootimg` and patch it
The current `mkbootimg` in trixie is unusable due to #1050300, so the
images won't build. In order to test those changes, it is therefore
necessary to patch it after install.

This commit will be dropped once #1050300 is fixed as (the now-working)
`mkbootimg` will then be pulled in directly through
`qcom-support-common`.

IMPORTANT: the resulting images MUST NOT be used for any other purpose
than testing.